Cleanup component for Msa Counter
